Rename `utils/utils.js` to `utils/index.js`
Allows importing the file with just `.../utils`
instead of `.../utils/utils`
