+++ utils/index.js
@@ -1,8 +1,6 @@
-const now = Date.now()
-
 const utils = {
   isDraft: (data) => data.page.filePathStem.includes('/drafts/'),
-  isScheduled: (data) => data.date >= now,
+  isScheduled: (data) => data.date >= Date.now(),
 }
 
 module.exports = utils
